refactor: use our own implementation of debounce
